Fix: retry connection to redis when lost
